+++ apps/c/CloverLeaf_3D_HDF5/OpenACC/calc_dt_kernel_get_openacc_kernel.cpp
@@ -95,29 +95,41 @@ void ops_par_loop_calc_dt_kernel_get(char const *name, ops_block block, int dim,
   #endif
 
   //set up initial pointers
-  int base0 = args[0].dat->base_offset + (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size) * start[0] * args[0].stencil->stride[0];
-  base0 = base0 + (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size) *
-    args[0].dat->size[0] *
-    start[1] * args[0].stencil->stride[1];
-  base0 = base0 + (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size) *
-    args[0].dat->size[0] *
-    args[0].dat->size[1] *
-    start[2] * args[0].stencil->stride[2];
-  #ifdef OPS_GPU
+  long long int base0 =
+      args[0].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[0].dat->type_size
+                                                     : args[0].dat->elem_size) *
+                                     start[0] * args[0].stencil->stride[0];
+  base0 = base0 +
+          (long long int)(block->instance->OPS_soa ? args[0].dat->type_size
+                                                   : args[0].dat->elem_size) *
+              args[0].dat->size[0] * start[1] * args[0].stencil->stride[1];
+  base0 = base0 + (long long int)(block->instance->OPS_soa
+                                      ? args[0].dat->type_size
+                                      : args[0].dat->elem_size) *
+                      args[0].dat->size[0] * args[0].dat->size[1] * start[2] *
+                      args[0].stencil->stride[2];
+#ifdef OPS_GPU
   double *p_a0 = (double *)((char *)args[0].data_d + base0);
   #else
   double *p_a0 = (double *)((char *)args[0].data + base0);
   #endif
 
-  int base1 = args[1].dat->base_offset + (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size) * start[0] * args[1].stencil->stride[0];
-  base1 = base1 + (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size) *
-    args[1].dat->size[0] *
-    start[1] * args[1].stencil->stride[1];
-  base1 = base1 + (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size) *
-    args[1].dat->size[0] *
-    args[1].dat->size[1] *
-    start[2] * args[1].stencil->stride[2];
-  #ifdef OPS_GPU
+  long long int base1 =
+      args[1].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[1].dat->type_size
+                                                     : args[1].dat->elem_size) *
+                                     start[0] * args[1].stencil->stride[0];
+  base1 = base1 +
+          (long long int)(block->instance->OPS_soa ? args[1].dat->type_size
+                                                   : args[1].dat->elem_size) *
+              args[1].dat->size[0] * start[1] * args[1].stencil->stride[1];
+  base1 = base1 + (long long int)(block->instance->OPS_soa
+                                      ? args[1].dat->type_size
+                                      : args[1].dat->elem_size) *
+                      args[1].dat->size[0] * args[1].dat->size[1] * start[2] *
+                      args[1].stencil->stride[2];
+#ifdef OPS_GPU
   double *p_a1 = (double *)((char *)args[1].data_d + base1);
   #else
   double *p_a1 = (double *)((char *)args[1].data + base1);
@@ -125,15 +137,21 @@ void ops_par_loop_calc_dt_kernel_get(char const *name, ops_block block, int dim,
 
   double *p_a2 = arg2h;
   double *p_a3 = arg3h;
-  int base4 = args[4].dat->base_offset + (block->instance->OPS_soa ? args[4].dat->type_size : args[4].dat->elem_size) * start[0] * args[4].stencil->stride[0];
-  base4 = base4 + (block->instance->OPS_soa ? args[4].dat->type_size : args[4].dat->elem_size) *
-    args[4].dat->size[0] *
-    start[1] * args[4].stencil->stride[1];
-  base4 = base4 + (block->instance->OPS_soa ? args[4].dat->type_size : args[4].dat->elem_size) *
-    args[4].dat->size[0] *
-    args[4].dat->size[1] *
-    start[2] * args[4].stencil->stride[2];
-  #ifdef OPS_GPU
+  long long int base4 =
+      args[4].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[4].dat->type_size
+                                                     : args[4].dat->elem_size) *
+                                     start[0] * args[4].stencil->stride[0];
+  base4 = base4 +
+          (long long int)(block->instance->OPS_soa ? args[4].dat->type_size
+                                                   : args[4].dat->elem_size) *
+              args[4].dat->size[0] * start[1] * args[4].stencil->stride[1];
+  base4 = base4 + (long long int)(block->instance->OPS_soa
+                                      ? args[4].dat->type_size
+                                      : args[4].dat->elem_size) *
+                      args[4].dat->size[0] * args[4].dat->size[1] * start[2] *
+                      args[4].stencil->stride[2];
+#ifdef OPS_GPU
   double *p_a4 = (double *)((char *)args[4].data_d + base4);
   #else
   double *p_a4 = (double *)((char *)args[4].data + base4);
